refactor: extract reusable ETS hierarchy import service
