+++ obs/api/v1/services/__init__.py
@@ -1,0 +1,1 @@
+"""Reusable service helpers for API v1 endpoints."""
